fix(configuration/host): operator< first compares host_id
Also, a function is added to look for a host by its name. Whereas
the key is an integer, the configuration is always written with
hostnames. So we must be able to look for a host by its hostname.

This function is not optimized, we must go through the set until
we find the host. One day, we will have services directly configured
with host id and this function will be removed.